milestone03: add controllers and make create in pokemonFacade use long.

+++ pkmnleague-api/src/main/java/cz/fi/muni/pa165/seminar/pkmnleague/facade/PokemonFacade.java
@@ -44,7 +44,7 @@ public interface PokemonFacade {
         * @param id
         * @return PokemonDTO
         */
-        public PokemonDTO getPokemonWithId(int id);
+        public PokemonDTO getPokemonWithId(long id);
         
         /**
         * Gets existing Pokemons by Trainer ID
